Move Sorted maps to CLDC.
