+++ modules/cldc-compact/src/main/java/cc/squirreljme/runtime/cldc/util/SortedTreeSet.java
@@ -8,7 +8,7 @@
 // See license.mkd for licensing and copyright information.
 // ---------------------------------------------------------------------------
 
-package net.multiphasicapps.collections;
+package cc.squirreljme.runtime.cldc.util;
 
 import java.util.AbstractSet;
 import java.util.Collection;
