+++ modules/cldc-compact/src/main/java/cc/squirreljme/runtime/cldc/util/__SortedTreeIterator__.java
@@ -8,7 +8,7 @@
 // See license.mkd for licensing and copyright information.
 // ---------------------------------------------------------------------------
 
-package net.multiphasicapps.collections;
+package cc.squirreljme.runtime.cldc.util;
 
 import java.util.ConcurrentModificationException;
 import java.util.Iterator;
